Fix intsimdmatrixneon.cpp stack corruption.
My understanding of the intsimdmatrix mechanism was that the
inputs would be resized so that we'd only ever get "whole
blocks" of data (i.e. that num_out would be a multiple of
kNumOutputsPerRegister, and num_in would be a multiple of
kNumInputsPerRegister). Sadly, this does not seem to be the
case, as we can get called (sometimes) with num_out % 8 == 7.

It's not clear to me WHY this happens. Is it a mistake in
the calling code? A quick investigation reveals that the
offending matrices are being read from the file and are not
obviously resized before being used.

Possibly we could benefit from resizing those matrices so
that special cases in this innermost loop are not actually
required, but unless and until that is done, let's fix the
inner loop.